Create Day 10. 11, 12 - Hash Maps.py
Day 10 and 11 were purely conceptual as I learned about the basics of a hash map/table and why it is a good concept/data structure. 
I also learned about collision and collision resolution techniques, such as open assignment and separate chaining. I learned that open assignment, which is defined as a technique in which a function places the collision bucket in the next available bucket, there were two main types: 1.  Linear probing 2. Quadratic probing.